Published another few types (UnicastRouter and IUnicastRouter)
