+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -130,9 +130,8 @@
     <Compile Include="Recoverability\Faults\MessageFaulted.cs" />
     <Compile Include="ReleaseDateAttribute.cs" />
     <Compile Include="Recoverability\ProcessingFailureInfo.cs" />
-    <Compile Include="Routing\IDistributionPolicy.cs" />
+    <Compile Include="Routing\IUnicastRouter.cs" />
     <Compile Include="Routing\RoutingMappingSettings.cs" />
-    <Compile Include="Routing\SpecificInstanceDistributionPolicy.cs" />
     <Compile Include="Routing\UnicastPublishRouter.cs" />
     <Compile Include="Routing\UnicastSendRouter.cs" />
     <Compile Include="Serialization\SerializationFeature.cs" />
@@ -264,7 +263,6 @@
     <Compile Include="Routing\AllInstancesDistributionStrategy.cs" />
     <Compile Include="Routing\UnicastRoute.cs" />
     <Compile Include="Routing\DistributionPolicy.cs" />
-    <Compile Include="Routing\IUnicastRouter.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\PublisherAddress.cs" />
     <Compile Include="Routing\SingleInstanceRoundRobinDistributionStrategy.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\SubscriptionRouter.cs" />
