--- src/NServiceBus.Core/Routing/IDistributionPolicy.cs
@@ -1,9 +1,0 @@
-namespace NServiceBus
-{
-    using Routing;
-
-    interface IDistributionPolicy
-    {
-        DistributionStrategy GetDistributionStrategy(string endpointName);
-    }
-}
